refactor: simplify pdftotext
This commit simplifies the pdftotext's interface making it more
composable. I use Turtle.FilePaths instead of Text to make the interface
safer and more informative. I also made the text output into an in-memory
data structure. The last change in some reduces composability when we
operate on large files, but allows callers to choose if they want to
interact with the filesystem.